MixTaskArchiveDeps v0.1.0 Mix.Tasks.Archive.Build.Elixir

Builds each elixir project required by the project into archives according to the specification of the Erlang Archive Format.

This task goal is to create packages. which could be used from Erlang environment without Elixir installation.

The archives will be created in the “archives” subdirectory of the project build directory by default, unless an argument -o is provided with the directory name.

The task will create archives for applications, referenced by the project and it’s dependencies, unless -a flag is specified.

Command line options

-o|--destination - specifies output directory name. Defaults to BUILD_DIR/archives

-a|--all_applications - specifies if all elixir applications should be archived. Defaults to false

Functions

build_archives(opts)

run(argv)

run(OptionParser.argv) :: :ok

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.